Move channel category setter to only those builders that support it.
