+++ javacord-api/src/main/java/org/javacord/api/entity/channel/ServerTextChannelBuilder.java
@@ -50,7 +50,12 @@ public class ServerTextChannelBuilder extends ServerChannelBuilder {
         return this;
     }
 
-    @Override
+    /**
+     * Sets the category of the channel.
+     *
+     * @param category The category of the channel.
+     * @return The current instance in order to chain call methods.
+     */
     public ServerTextChannelBuilder setCategory(ChannelCategory category) {
         delegate.setCategory(category);
         return this;
