+++ javacord-api/src/main/java/org/javacord/api/entity/channel/ServerVoiceChannelBuilder.java
@@ -39,7 +39,13 @@ public class ServerVoiceChannelBuilder extends ServerChannelBuilder {
         return this;
     }
 
-    @Override
+
+    /**
+     * Sets the category of the channel.
+     *
+     * @param category The category of the channel.
+     * @return The current instance in order to chain call methods.
+     */
     public ServerVoiceChannelBuilder setCategory(ChannelCategory category) {
         delegate.setCategory(category);
         return this;
