+++ javacord-core/src/main/java/org/javacord/core/entity/channel/ServerTextChannelBuilderDelegateImpl.java
@@ -2,6 +2,7 @@ package org.javacord.core.entity.channel;
 
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import org.javacord.api.entity.channel.ChannelCategory;
 import org.javacord.api.entity.channel.ServerTextChannel;
 import org.javacord.api.entity.channel.internal.ServerTextChannelBuilderDelegate;
 import org.javacord.core.entity.server.ServerImpl;
@@ -23,6 +24,11 @@ public class ServerTextChannelBuilderDelegateImpl extends ServerChannelBuilderDe
     private String topic = null;
 
     /**
+     * The category of the channel.
+     */
+    private ChannelCategory category = null;
+
+    /**
      * Creates a new server text channel builder delegate.
      *
      * @param server The server of the server text channel.
@@ -37,12 +43,20 @@ public class ServerTextChannelBuilderDelegateImpl extends ServerChannelBuilderDe
     }
 
     @Override
+    public void setCategory(ChannelCategory category) {
+        this.category = category;
+    }
+
+    @Override
     public CompletableFuture<ServerTextChannel> create() {
         ObjectNode body = JsonNodeFactory.instance.objectNode();
         body.put("type", 0);
         super.prepareBody(body);
         if (topic != null) {
             body.put("topic", topic);
+        }
+        if (category != null) {
+            body.put("parent_id", category.getIdAsString());
         }
         return new RestRequest<ServerTextChannel>(server.getApi(), RestMethod.POST, RestEndpoint.SERVER_CHANNEL)
                 .setUrlParameters(server.getIdAsString())
